Launcher3: Search bar improvements
This is more UI/UX changes to make it more compatible with our various search engines.

Let's change the icons according to the search engine you pick to use.
Also picking @shagbag913 search hint toggle because it might be an annoyance for some, to others it's nice to have the text there.
* Bing icon from: materialdesignicons.com
* Duckduckgo icon from: icons8.com

Co-authored-by: shagbag913 <sh4gbag913@gmail.com>